requirements.txt
browser-use>=0.11.9
playwright>=1.40.0
pydantic>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
openai>=1.0.0
pymongo>=4.6.0
typing-extensions>=4.8.0
lmnr>=0.1.0
PyPDF2>=3.0.0
apscheduler>=3.10.0
pyyaml>=6.0
gspread>=6.0.0
google-auth>=2.0.0
